// packages/sdk/src/auth.ts
export interface AuthTokens {
  tokenExpiry: number | undefined;
  refreshToken: string | undefined;
  accessToken: string | undefined;
}